# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Luminox.",
        "title": "Luminox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/space/{space_id}/block": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List blocks in a space. Use type query parameter to filter by block type (page, folder, sop). Use parent_id query parameter to filter by parent. If both type and parent_id are empty, returns top-level pages and folders.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "List blocks",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "page",
                            "folder",
                            "sop"
                        ],
                        "type": "string",
                        "description": "Block type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Parent ID",
                        "name": "parent_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Block"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# List blocks\nblocks = client.blocks.list(\n    space_id='space-uuid',\n    parent_id='parent-uuid',\n    block_type='page'\n)\nfor block in blocks:\n    print(f\"{block.id}: {block.title}\")\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// List blocks\nconst blocks = await client.blocks.list('space-uuid', {\n  parentId: 'parent-uuid',\n  type: 'page'\n});\nfor (const block of blocks) {\n  console.log(`${block.id}: ${block.title}`);\n}\n"
                    }
                ]
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new block (supports types: page, folder, sop). For page and folder types, parent_id is optional. For sop type, parent_id is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Create block",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CreateBlock payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateBlockReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpclient.InsertBlockResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Create a folder\nfolder = client.blocks.create(\n    space_id='space-uuid',\n    block_type='folder',\n    title='My Folder'\n)\n\n# Create a page under the folder\npage = client.blocks.create(\n    space_id='space-uuid',\n    parent_id=folder['id'],\n    block_type='page',\n    title='My Page',\n    props={\"description\": \"Page content here\"}\n)\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Create a folder\nconst folder = await client.blocks.create('space-uuid', {\n  blockType: 'folder',\n  title: 'My Folder'\n});\n\n// Create a page under the folder\nconst page = await client.blocks.create('space-uuid', {\n  parentId: folder.id,\n  blockType: 'page',\n  title: 'My Page',\n  props: { description: 'Page content here' }\n});\n"
                    }
                ]
            }
        },
        "/space/{space_id}/block/{block_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a block by its ID (works for block types: page, folder, sop)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Delete block",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Block ID",
                        "name": "block_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Delete a block\nclient.blocks.delete(space_id='space-uuid', block_id='block-uuid')\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Delete a block\nawait client.blocks.delete('space-uuid', 'block-uuid');\n"
                    }
                ]
            }
        },
        "/space/{space_id}/block/{block_id}/move": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move block by updating its parent_id. Works for block types: page, folder, sop. For page and folder types, parent_id can be null (root level).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Move block",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Block ID",
                        "name": "block_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MoveBlock payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MoveBlockReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Move block to a different parent\nclient.blocks.move(\n    space_id='space-uuid',\n    block_id='block-uuid',\n    parent_id='new-parent-uuid'\n)\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Move block to a different parent\nawait client.blocks.move('space-uuid', 'block-uuid', {\n  parentId: 'new-parent-uuid'\n});\n"
                    }
                ]
            }
        },
...
        "/space/{space_id}/block/{block_id}/sort": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update block sort value (works for block types: page, folder, sop)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Update block sort",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Block ID",
                        "name": "block_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateBlockSort payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateBlockSortReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Update block sort order\nclient.blocks.update_sort(\n    space_id='space-uuid',\n    block_id='block-uuid',\n    sort=5\n)\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Update block sort order\nawait client.blocks.updateSort('space-uuid', 'block-uuid', {\n  sort: 5\n});\n"
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateBlockReq": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "parent_id": {
                    "type": "string"
                },
                "props": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "page"
                }
            }
        },
...
        "handler.MoveBlockReq": {
            "type": "object",
            "properties": {
                "parent_id": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateBlockSortReq": {
            "type": "object",
            "properties": {
                "sort": {
                    "type": "integer"
                }
            }
        },
...
        "httpclient.InsertBlockResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "model.Block": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "parent_id": {
                    "type": "string"
                },
                "props": {
                    "type": "object"
                },
                "sort": {
                    "type": "integer"
                },
                "space_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}